Add tracking and name properties to Rental and User classes
The `Rental` class in the `Karent.DataModel` namespace has been
updated to include two new properties: `ModifiedBy` and `ModifiedOn`.
These properties are intended to track the user who last modified
the rental record and the date and time when the modification occurred.

The `User` class in the `Karent.DataModel` namespace has been updated
to include a new property: `Name`. This property is intended to store
the name of the user.